types: Moved VideoBackend from core to types
