+++ types/Cargo.toml
@@ -18,6 +18,7 @@ png = { version = "0.17.5" }
 gif = "0.11.4"
 downcast-rs = "1.2.0"
 flate2 = "1.0.24"
+generational-arena = "0.2.8"
 
 [dependencies.jpeg-decoder]
 version = "0.2.6"
