+++ types/src/backend.rs
@@ -2,3 +2,4 @@ pub mod log;
 pub mod render;
 pub mod storage;
 pub mod ui;
+pub mod video;
